A search box that can filter tilesets (#10)
* Add ControlAutoUnfocus

ControlAutoUnfocus is a helper node that automatically releases focus when the user leaves, e.g. clicked somewhere. The main use is, especially for LineEdit, which has a problem when the focus is always there even when the user clicked outside of it.

* Add magnifying icon for search box

* Refactor inspector panel's node structure

* Add tileset filtering

A working search box that can filter tilesets